services/ais_ingestion/requirements.txt
websockets>=10.0
python-dotenv>=0.19.0
neo4j>=5.14.0
